Finished cleanup bullet points
