+++ src/web/command_control/client/style/CommandControl.less
@@ -532,7 +532,7 @@ td {
     background: @overlay-background;
     border-radius: 0 @panel-corner-radius 0 0;
     display: flex;
-    flex-flow: column-reverse wrap-reverse;
+    flex-flow: column wrap;
   }
 
   #measureResult {
